Bug 801: Graph TextRenderer Cleanup Part-2: Remove Path2D from Glyph/Typecast* ; Misc Cleanup
Commit c3621221b9a563495b4f54fe60e18e8db8cc57fb introduced
create an OutlineShape per Glyph from it's data w/o going through Path2D.

Misc Cleanup: Remove unused code/fields, use private/final where possible.

+++ src/jogl/classes/jogamp/graph/font/typecast/TypecastHMetrics.java
@@ -50,7 +50,7 @@ class TypecastHMetrics implements Metrics {
         headTable = this.fontImpl.font.getHeadTable();
         hheaTable = this.fontImpl.font.getHheaTable();
         // vheaTable = this.fontImpl.font.getVheaTable();
-        unitsPerEM_Inv = 1.0f / ( (float) headTable.getUnitsPerEm() );
+        unitsPerEM_Inv = 1.0f / ( headTable.getUnitsPerEm() );
 
         int maxWidth = headTable.getXMax() - headTable.getXMin();
         int maxHeight = headTable.getYMax() - headTable.getYMin();
